requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# METALoci/build/lib/metaloci/plot/plot.py: 13
# METALoci/build/lib/metaloci/tools/figure.py: 15
# METALoci/metaloci/plot/plot.py: 13
# METALoci/metaloci/tools/figure.py: 15
Pillow == 10.0.0

# METALoci/build/lib/metaloci/misc/misc.py: 9
# METALoci/build/lib/metaloci/tools/layout.py: 11
# METALoci/metaloci/misc/misc.py: 9
# METALoci/metaloci/tools/layout.py: 11
cooler == 0.9.2

# METALoci/build/lib/metaloci/mlo.py: 1
# METALoci/metaloci/mlo.py: 1
dill == 0.3.6

# METALoci/build/lib/metaloci/spatial_stats/lmi.py: 11
# METALoci/metaloci/spatial_stats/lmi.py: 11
esda == 2.4.3

# METALoci/build/lib/metaloci/plot/plot.py: 3
# METALoci/build/lib/metaloci/spatial_stats/lmi.py: 7
# METALoci/build/lib/metaloci/tools/figure.py: 12
# METALoci/metaloci/plot/plot.py: 3
# METALoci/metaloci/spatial_stats/lmi.py: 7
# METALoci/metaloci/tools/figure.py: 12
geopandas == 0.13.2

# METALoci/build/lib/metaloci/misc/misc.py: 11
# METALoci/build/lib/metaloci/tools/layout.py: 12
# METALoci/metaloci/misc/misc.py: 11
# METALoci/metaloci/tools/layout.py: 12
hic_straw == 1.3.1

# METALoci/build/lib/metaloci/plot/plot.py: 4
# METALoci/build/lib/metaloci/spatial_stats/lmi.py: 8,12
# METALoci/metaloci/plot/plot.py: 4
# METALoci/metaloci/spatial_stats/lmi.py: 8,12
libpysal == 4.6.2

# METALoci/build/lib/metaloci/plot/plot.py: 12
# METALoci/build/lib/metaloci/tools/figure.py: 13
# METALoci/build/lib/metaloci/tools/layout.py: 13
# METALoci/metaloci/plot/plot.py: 12
# METALoci/metaloci/tools/figure.py: 13
# METALoci/metaloci/tools/layout.py: 13
matplotlib == 3.7.2

# METALoci/build/lib/metaloci/plot/plot.py: 6
# METALoci/build/lib/metaloci/tools/layout.py: 14
# METALoci/metaloci/plot/plot.py: 6
# METALoci/metaloci/tools/layout.py: 14
networkx == 3.1

# METALoci/build/lib/metaloci/tools/figure.py: 16
# METALoci/build/lib/metaloci/tools/ml.py: 16
# METALoci/build/lib/metaloci/tools/prep.py: 14
# METALoci/metaloci/tools/figure.py: 16
# METALoci/metaloci/tools/ml.py: 16
# METALoci/metaloci/tools/prep.py: 14
numba == 0.55.1

# METALoci/build/lib/metaloci/graph_layout/kk.py: 1
# METALoci/build/lib/metaloci/misc/misc.py: 1
# METALoci/build/lib/metaloci/plot/plot.py: 7
# METALoci/build/lib/metaloci/spatial_stats/lmi.py: 9
# METALoci/metaloci/graph_layout/kk.py: 1
# METALoci/metaloci/misc/misc.py: 1
# METALoci/metaloci/plot/plot.py: 7
# METALoci/metaloci/spatial_stats/lmi.py: 9
numpy == 1.21.6

# METALoci/build/lib/metaloci/misc/misc.py: 2
# METALoci/build/lib/metaloci/plot/plot.py: 8
# METALoci/build/lib/metaloci/spatial_stats/lmi.py: 10
# METALoci/build/lib/metaloci/tools/figure.py: 14
# METALoci/build/lib/metaloci/tools/layout.py: 15
# METALoci/build/lib/metaloci/tools/ml.py: 13
# METALoci/build/lib/metaloci/tools/prep.py: 11
# METALoci/build/lib/metaloci/utility_scripts/prep_coords_genes.py: 4
# METALoci/metaloci/misc/misc.py: 2
# METALoci/metaloci/plot/plot.py: 8
# METALoci/metaloci/spatial_stats/lmi.py: 10
# METALoci/metaloci/tools/figure.py: 14
# METALoci/metaloci/tools/layout.py: 15
# METALoci/metaloci/tools/ml.py: 13
# METALoci/metaloci/tools/prep.py: 11
# METALoci/metaloci/utility_scripts/prep_coords_genes.py: 4
pandas == 1.3.0

# METALoci/build/lib/metaloci/tools/prep.py: 10
# METALoci/metaloci/tools/prep.py: 10
pybedtools == 0.9.0

# METALoci/build/lib/metaloci/plot/plot.py: 15
# METALoci/build/lib/metaloci/spatial_stats/lmi.py: 13
# METALoci/build/lib/metaloci/tools/layout.py: 16,17
# METALoci/metaloci/plot/plot.py: 15
# METALoci/metaloci/spatial_stats/lmi.py: 13
# METALoci/metaloci/tools/layout.py: 16,17
scipy == 1.9.1

# METALoci/build/lib/metaloci/plot/plot.py: 9
# METALoci/metaloci/plot/plot.py: 9
seaborn == 0.11.0

# METALoci/setup.py: 4
setuptools == 59.8.0

# METALoci/build/lib/metaloci/plot/plot.py: 16,17
# METALoci/build/lib/metaloci/spatial_stats/lmi.py: 15,16
# METALoci/metaloci/plot/plot.py: 16,17
# METALoci/metaloci/spatial_stats/lmi.py: 15,16
shapely == 2.0.1

# METALoci/build/lib/metaloci/misc/misc.py: 3
# METALoci/build/lib/metaloci/utility_scripts/prep_coords_genes.py: 10
# METALoci/metaloci/misc/misc.py: 3
# METALoci/metaloci/utility_scripts/prep_coords_genes.py: 10
tqdm == 4.65.0
